Rename targets and move around some code
bla
