+++ Sources/NIOCertificateReloading/TimedCertificateReloader.swift
@@ -26,20 +26,6 @@ import FoundationEssentials
 import Foundation
 #endif
 
-/// A protocol that defines a certificate reloader.
-///
-/// A certificate reloader is a service that can provide you with updated versions of a certificate and private key pair, in
-/// the form of a `NIOSSLContextConfigurationOverride`, which will be used when performing a TLS handshake in NIO.
-/// Each implementation can choose how to observe for changes, but they all require an ``sslContextConfigurationOverride``
-/// to be exposed.
-@available(macOS 11.0, iOS 14, tvOS 14, watchOS 7, *)
-public protocol CertificateReloader: Sendable {
-    /// A `NIOSSLContextConfigurationOverride` that will be used as part of the NIO application's TLS configuration.
-    /// Its certificate and private key will be kept up-to-date via whatever mechanism the specific ``CertificateReloader``
-    /// implementation provides.
-    var sslContextConfigurationOverride: NIOSSLContextConfigurationOverride { get async }
-}
-
 /// A ``TimedCertificateReloader`` is an implementation of a ``CertificateReloader``, where the certificate and private
 /// key pair is updated at a fixed interval from the file path or memory location configured.
 ///
@@ -218,10 +204,10 @@ public struct TimedCertificateReloader: CertificateReloader {
 
     private func reloadPair() {
         if let certificateBytes = self.loadCertificate(),
-            let keyBytes = self.loadPrivateKey(),
-            let certificate = self.parseCertificate(from: certificateBytes),
-            let key = self.parsePrivateKey(from: keyBytes),
-            key.publicKey.isValidSignature(certificate.signature, for: certificate)
+           let keyBytes = self.loadPrivateKey(),
+           let certificate = self.parseCertificate(from: certificateBytes),
+           let key = self.parsePrivateKey(from: keyBytes),
+           key.publicKey.isValidSignature(certificate.signature, for: certificate)
         {
             self.attemptToUpdatePair(certificate: certificate, key: key)
         }
@@ -299,19 +285,6 @@ public struct TimedCertificateReloader: CertificateReloader {
                 privateKey: .privateKey(nioSSLPrivateKey)
             )
         }
-    }
-}
-
-extension TLSConfiguration {
-    /// Configure a ``CertificateReloader`` to observe updates for the certificate and key pair used.
-    /// - Parameter reloader: A ``CertificateReloader`` to watch for certificate and key pair updates.
-    /// - Returns: A `TLSConfiguration` that reloads the certificate and key used in its SSL handshake.
-    @available(macOS 11.0, iOS 14, tvOS 14, watchOS 7, *)
-    mutating public func setCertificateReloader(_ reloader: some CertificateReloader) -> Self {
-        self.sslContextCallback = { _, promise in
-            promise.completeWithTask { await reloader.sslContextConfigurationOverride }
-        }
-        return self
     }
 }
